Actually implement Linux getmntent support
To make testing more convenient, this build adds a Dockerfile which runs
it on a Debian image with syslogd available since that's currently a
hard requirement.

Since we're just using this to run the daemon it's started with a
timeout long enough to run for one iteration. It might be useful to
consider the testing strategy at some point for confirming output to
syslog and the pushgateway but in the meantime we're just dumping
everything to the console.