Refactor get_mount_points implementations to return Result
@Dr-Emann flagged an incorrect test for getmntinfo()
returning an error – which is a rather unlikely but possible situation – and made the suggestion that the platform-specific implementations could use Result.

This changes the return type to Result<Vec<PathBuf>>
and in the BSD implementation will return the error from io::Error::last_os_error() — i.e. errno on Unix — if getmntinfo() returns 0.